refactor: 패키지 구조 변경 및 관련 코드 정리
- 필요 없는 테스트 메서드 코드 제거
