+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/handler/mapping/HandlerKey.java
@@ -1,7 +1,6 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc.tobe.handler.mapping;
 
 import com.interface21.web.bind.annotation.RequestMethod;
-
 import java.util.Objects;
 
 public class HandlerKey {
@@ -24,8 +23,12 @@ public class HandlerKey {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof HandlerKey)) return false;
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof HandlerKey)) {
+            return false;
+        }
         HandlerKey that = (HandlerKey) o;
         return Objects.equals(url, that.url) && requestMethod == that.requestMethod;
     }
